+++ lib/glimmer/swt/custom/shape/rectangle.rb
@@ -84,11 +84,6 @@ module Glimmer
             [[x, y], [x + width, y], [x + width, y + height], [x, y + height]]
           end
           
-          # checks if shape contains the point denoted by x and y
-          def contain?(x, y)
-            x.between?(self.x, self.x + width) && y.between?(self.y, self.y + height)
-          end
-          
           # checks if drawn or filled rectangle includes the point denoted by x and y (if drawn, it only returns true if point lies on the edge)
           def include?(x, y)
             if filled?